// apps/website/src/types/splint.ts
export interface SplintData {
    id: string;
    title: string;
    description: string;
    availability: string;
    "availability date": string;
    "expiration date": string;
    link: string;
    "mobile link": string;
    "image link": string;
    price: string;
    "sale price": string;
    "sale price effective date": string;
    "identifier exists": string;
    gtin: string;
    mpn: string;
    brand: string;
    "product highlight": string;
    "product detail": string;
    "additional image link": string;
    condition: string;
    adult: string;
    color: string;
    size: string;
    "size type": string;
    "size system": string;
    gender: string;
    material: string;
    pattern: string;
    "age group": string;
    multipack: string;
    "is bundle": string;
    "unit pricing measure": string;
    "unit pricing base measure": string;
    "energy efficiency class": string;
    "min energy efficiency class": string;
    "item group id": string;
    "sell on google quantity": string;
}